Create code
This function takes an optional argument start_time which represents the starting time of the stopwatch. If start_time is not provided (i.e., it's None), the function returns 0. Otherwise, it returns the elapsed time since start_time in seconds, using the time.monotonic() function to get the current time.

Note that this function has no side-effects, and it doesn't modify any external state. It simply takes an input and returns an output, making it a pure function.